+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic
 Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+## Changed
+- Update to use sensu-plugin-sdk 0.16
 
 ## [0.9.0] - 2021-01-27